doit: Split out cmd_console
